feat: Add Skin DamageLevel support
Retrieving DamageStates of skins is surprisingly difficult. Most methods that set and load these states are native and hence harder to decompile.
A working solution will surely have to use GRI as it's the object responsible for saved data.
The game internally sets the "BatmanDamageLevelx" global flags to false when loading in Challenge Maps it seems.
This is why the current method reads these flags but it doesn't seem to work. And there's also the question of Catwoman support. I'll have to revisit this later on.