feat: Add DTO for RPCs and sync after swap
When switching characters, the game resets a lot of attributes like health, animations and camera position. Consequently, It is important to store those values and reapply them after switching.
I started off with a few variables in SwapCharachter() but that quickly became extremely inconsistent and hard to manage.
Thus, I opted for the PlayerState record with its helper functions to save and apply snapshots of RPCs.