feat: Spawn smoke bomb particle on switch
SSG suggested that a Ninja smoke bomb would be a good transition effect between characters. It would also hide some inconsistencies on character switch. So I added it.
When the engine first becomes available (`#Main()`), I load in the Under_C2_Ch5 UPK, prevent it from being GC'd and save the particle template once for the entire duration of the session.
That package is the smallest one containing the effect (1.3 MB) but I could create an SF UPK for it. Alternatively, I wanted to load in "Playable_Robin_SF" anyway (when I figure out how to keep him in RAM forever). That would improve performance whilst switching characters and would load in the effect for me.
Lastly, I instantiate a particle with the template and attach it to the RPP. This isn't ideal, however, because the smoke artificially follows the player. I haven't found a way to spawn in the particle in the world.