fix: Detective mode sync inconsistent with Bruce
When you had detective vision on and switched to Bruce Wayne, the script would manually reenable bInvestigateVision.
That doesn't do much while playing as Bruce because the game checks every tick if the RPC has a forensics device. If not (Bruce), no filter is rendered.
However, when switching again, the mod would automatically apply detective vision although it was visually turned off to the player during the time they played as Bruce Wayne.
That's because the attribute was never changed by the game automatically and detective mode toggles are ignored when no forensics device is present.